Add Bootstrap component initialization functions for dropdown, collapse, and tooltip
